[DX] Extract JobInterruptedException and JobRepositoryInterface
