+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Job/JobInterruptedExceptionTest.php
@@ -3,7 +3,7 @@
 namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Job;
 
 use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
-use Akeneo\Bundle\BatchBundle\Job\JobInterruptedException;
+use Akeneo\Component\Batch\Job\JobInterruptedException;
 
 /**
  * Tests related to the JobInterruptedException
