+++ src/Akeneo/Component/Batch/Step/AbstractStep.php
@@ -8,8 +8,8 @@ use Akeneo\Component\Batch\Event\InvalidItemEvent;
 use Akeneo\Component\Batch\Event\StepExecutionEvent;
 use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
 use Akeneo\Bundle\BatchBundle\Job\ExitStatus;
-use Akeneo\Bundle\BatchBundle\Job\JobInterruptedException;
-use Akeneo\Bundle\BatchBundle\Job\JobRepositoryInterface;
+use Akeneo\Component\Batch\Job\JobInterruptedException;
+use Akeneo\Component\Batch\Job\JobRepositoryInterface;
 use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
@@ -65,7 +65,7 @@ abstract class AbstractStep implements StepInterface
     /**
      * Public setter for {@link JobRepositoryInterface}.
      *
-     * @param JobRepositoryInterface $jobRepository jobRepository is a mandatory dependence (no default).
+     * @param \Akeneo\Component\Batch\Job\JobRepositoryInterface $jobRepository jobRepository is a mandatory dependence (no default).
      *
      * @return AbstractStep
      */
