+++ src/Pim/Bundle/ConnectorBundle/JobLauncher/SimpleJobLauncher.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\ConnectorBundle\JobLauncher;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
-use Akeneo\Bundle\BatchBundle\Job\JobRepositoryInterface;
+use Akeneo\Component\Batch\Job\JobRepositoryInterface;
 use Akeneo\Bundle\BatchBundle\Launcher\SimpleJobLauncher as BaseSimpleJobLauncher;
 use Doctrine\Common\Persistence\ObjectManager;
 use Pim\Component\Connector\Factory\JobConfigurationFactory;
@@ -27,7 +27,7 @@ class SimpleJobLauncher extends BaseSimpleJobLauncher
     protected $objectManager;
 
     /**
-     * @param JobRepositoryInterface  $jobRepository
+     * @param \Akeneo\Component\Batch\Job\JobRepositoryInterface  $jobRepository
      * @param JobConfigurationFactory $jobConfigFactory
      * @param ObjectManager           $objectManager
      * @param string                  $rootDir
